refactor: move searchProducts tests to dedicated file
